+++ HTMLStyler/HTMLStyler/Extensions/NSMutableAttributedString+trim.swift
